Pass all props to webview
Pass all props to the WebView before setting the specific WebView props required by react-native-remote-svg.

The `baseUrl` added to the `source` prop fixed an issue where iOS does not fire the `onLoad...` events. This also means that the calls to `onLoadStart` and 'onLoadEnd' can be removed, as they will be delegated to the WebView.